reorg cli add "add" commands to CLI
