+++ housekeeper/cli/core.py
@@ -1,14 +1,10 @@
 # -*- coding: utf-8 -*-
-import datetime
-
 import click
 import coloredlogs
 import ruamel.yaml
 
 import housekeeper
-from housekeeper.exc import VersionIncludedError
-from housekeeper.include import include_version
-from housekeeper.store import Store
+from . import add, include, init
 
 
 @click.group()
@@ -26,54 +22,6 @@ def base(context, config, database, root, log_level):
     context.obj['root'] = root if root else context.obj['root']
 
 
-@base.command()
-@click.option('--reset', is_flag=True, help='reset database before setting up tables')
-@click.option('--force', is_flag=True, help='bypass manual confirmations')
-@click.pass_context
-def init(context, reset, force):
-    """Setup the database."""
-    store = Store(context.obj['database'], context.obj['root'])
-    existing_tables = store.engine.table_names()
-    if force or reset:
-        if existing_tables and not force:
-            message = f"Delete existing tables? [{', '.join(existing_tables)}]"
-            click.confirm(click.style(message, fg='yellow'), abort=True)
-        store.drop_all()
-    elif existing_tables:
-        click.echo(click.style("Database already exists, use '--reset'", fg='red'))
-        context.abort()
-
-    store.create_all()
-    message = f"Success! New tables: {', '.join(store.engine.table_names())}"
-    click.echo(click.style(message, fg='green'))
-
-
-@base.command()
-@click.option('-v', '--version', type=int, help='version id of the bundle version')
-@click.argument('bundle_name', required=False)
-@click.pass_context
-def include(context, bundle_name, version):
-    """Include a bundle of files into the internal space.
-
-    Use bundle name if you simply want to inlcude the latest version.
-    """
-    store = Store(context.obj['database'], context.obj['root'])
-    if version:
-        version_obj = store.Version.get(version)
-        if version_obj is None:
-            click.echo(click.style('version not found', fg='red'))
-    else:
-        bundle_obj = store.bundle(bundle_name)
-        if bundle_obj is None:
-            click.echo(click.style('bundle not found', fg='red'))
-        version_obj = bundle_obj.versions[0]
-
-    try:
-        include_version(context.obj['root'], version_obj)
-    except VersionIncludedError as error:
-        click.echo(click.style(error.message, fg='red'))
-        context.abort()
-
-    version_obj.included_at = datetime.datetime.now()
-    store.commit()
-    click.echo(click.style('included all files!', fg='green'))
+base.add_command(init.init)
+base.add_command(add.add)
+base.add_command(include.include)
